+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Use `Role.permissions_names` instead of `Role.permissions`.
+
 ## [3.21.0] - 2025-03-24
 
 ### Changed